Move FadeHeaderBehavior to new base class
Was much more aligned then it looked at first glance. Think it was first, so some things were just done locally in scope vs. cached from the other classes. At least now we can clean things up all together easily if we want to optimize things.

Main difference is some of the event hooks, so if they're not needed we may want flags/options for those in the base class to enable.