fix: check constraints before allowing the resize to proceed
Fixes #339
If the vector contains another item with Star sizing, it's not enough
to just change current. The change will flow to the Star sized item
and not to the sibling if the sibling is fixed-size. We need to explicitly
apply the change to the sibling.
Now also checks the constraints before allowing the resize.