Use recursive wildcard for discovering nuget packages.
It should have output the *.nupkg for all packages to the path ./ we passed into the build script, but the ./*.nupkg wildcard failed to find them for artifact upload.

Either our script copies nupkg to the given folder locally but not in CI, or `dotnet nuget push ./*.nupkg` failed to locate nupkg files output to the working directory. Let's see which one.